Building final project combining with mlflow-server, minio, api-endpoint, and ui

+++ src/api/schemas.py
@@ -13,7 +13,7 @@ class CarFeatures(BaseModel):
     fuelType: Optional[str] = None
     Make: Optional[str] = None
 
-    year: Optional[int] = None
+    year: Optional[float] = None
 
     @model_validator(mode="after")
     def check_null_limit(self):
